+++ lib/conjur/cli.rb
@@ -49,6 +49,8 @@ module Conjur
   class CLI
     extend GLI::App
 
+    @current_command = nil
+
     class << self
       def load_config
         Conjur::Config.load
@@ -91,6 +93,19 @@ module Conjur
         load_plugins
         commands_from 'conjur/command'
       end
+
+      def appliance_version
+        Conjur::API.service_version 'appliance'
+      rescue
+        nil
+      end
+
+      def command_version_compatible? command
+        !command.instance_variable_defined?(:@conjur_min_version) ||
+          (appliance_version &&
+           command.instance_variable_get(:@conjur_min_version) <= appliance_version
+          )
+      end
     end
 
     init!
@@ -122,13 +137,22 @@ module Conjur
       
       true
     end
+
+    around do |global,command,options,args,code|
+      @current_command = command
+      code.call
+      @current_command = nil
+    end
     
     on_error do |exception|
       require 'rest-client'
       require 'patches/conjur/error'
-
+        
       run_default_handler = true
-      if exception.is_a?(RestClient::Exception) && exception.response
+      if @current_command != nil && !command_version_compatible?(@current_command)
+        $stderr.puts "error: this command is not supported by the current Conjur server version"
+        run_default_handler = false
+      elsif exception.is_a?(RestClient::Exception) && exception.response
         err = Conjur::Error.create exception.response.body
         if err
           $stderr.puts "error: " + err.message